[C10k] fix receiver side unexpected exit issue when test with 10K TCP connections
If test with 10K TCP connections, for example, receiver run with -P 500,
and sender side run with -P 500 -n 20, the receiver side may not abe to
read /proc files. In this case, the receiver side will exit with
"EXIT_FAILURE"
which is not a expected behavior.

This issue is fixed in this commit.