receiver: zero out the counter before a new test starting
We need to reset server side perf counters at the begining, after
light-is-on;
this is to handle these cases when:
a) receiver in sync mode, but sender connected as no_sync mode;
in this case, before light-is-on, the threads have some data counted
already.
b) receiver is running in a loop; the previous test has finished but the
sockets are still working and receiving data (data arrived with
latency);
so we need to reset the counter before a new test starting.