Add simple value check for options "-K" and "-I"
If runnning a test with mandotory options such as "-K" or "-I", but without a value provided, the option then will consume the next option as its value.

For example: "ntttcp -K -V": ntttcp will put "-V" as a value for "-K", which is incorrect.

This patch fixed this issue.