Change the test bandwidth limit implementation
Now implement the bandwidth limit by hold-on/pause the traffic in ntttcp application.
If the bandwidth is higher than the limit, then hold on senders to send traffic.

This bandwidth limit works for both TCP and UDP tests.

Future work: maybe we can increase the accuracy of the limit by:
a) more frequent throughput check;
b) or, set a range (high_water and low_water) to pause/resume the traffic.